Update functions to work with Netlify (aws)
