--- functions/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "functions",
-  "description": "Cloud Functions for Firebase",
-  "scripts": {
-    "lint": "eslint .",
-    "serve": "firebase serve --only functions",
-    "shell": "firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "dependencies": {
-    "cors": "^2.8.5",
-    "firebase-admin": "^6.5.0",
-    "firebase-functions": "^2.1.0"
-  },
-  "devDependencies": {},
-  "private": true
-}
